pyproject.toml
[project]
name = "halfvio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.7",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "ruff>=0.14.4",
    "ty>=0.0.1a25",
]
